fix: KM-50817 prevent k8s watch subscriptions from ending prematurely
grpc.experimental.Resolver#destroy does not mean that the resolver is
permanently destroyed. It can be revived by a call to #updateResolution.
Changed our handling of addListener/removeListener to account for this.

Additionally, moved k8s-watch's log messages to its own category since
they are very noisy and aren't as useful as k8s-resolver's log messages.